Added module description & removing invisible attribute of attr val lines + bump version
