+++ product_configurator_mrp/__manifest__.py
@@ -1,6 +1,6 @@
 {
     'name': 'Product Configurator Manufacturing',
-    'version': '11.0.1.0.0',
+    'version': '11.0.1.0.1',
     'category': 'Manufacturing',
     'summary': 'BOM Support for configurable products',
     'author': 'Pledra',
@@ -10,12 +10,11 @@
     "data": [
         'security/configurator_security.xml',
         'security/ir.model.access.csv',
-        'views/product_view.xml',
+        'views/product_attribute_view.xml',
         'views/product_config_view.xml',
         'views/mrp_view.xml',
     ],
     'demo': [
-        'demo/product_template.xml',
     ],
     'test': [],
     'installable': True,
