+++ product_configurator_mrp/models/__init__.py
@@ -1,3 +1,4 @@
 from . import product
+from . import product_attribute
 from . import product_config
 from . import mrp
