+++ product_configurator_mrp/views/product_attribute_view.xml
@@ -13,18 +13,4 @@
         </field>
     </record>
 
-            <!-- Add quantity to attribute lines -->
-
-<!--             <xpath expr="//field[@name='attribute_line_ids']//field[@name='custom']" position="after">
-    <field name="quantity" position="after"/>
-</xpath> -->
-
-            <!-- MRP Routing -->
-
-<!--             <xpath expr="//field[@name='config_image_ids']" position="after">
-                <separator name="configurator_mrp" string="MRP"/>
-                <group>
-                    <field name="routing_id"/>
-                </group>
-            </xpath> -->
 </odoo>
